Title Screen
Added a title screen scene. This will allow the player to start a new game with no save data or continue from their most recent save. They can also close the program from here. Currently has support for loading the living room fresh and should be able to close. Need to implement a method for reaching this screen in-game.